+++ packages/keyring-controller/src/index.ts
@@ -25,6 +25,7 @@ export type {
   KeyringControllerWithKeyringUnsafeAction,
   KeyringControllerWithKeyringV2Action,
   KeyringControllerWithKeyringV2UnsafeAction,
+  KeyringControllerExportSeedPhraseAction,
 } from './KeyringController-method-action-types';
 export type * from './types';
 export * from './errors';