+++ packages/transaction-controller/src/utils/gas.ts
@@ -23,7 +23,11 @@ import type {
   TransactionMeta,
   TransactionParams,
 } from '../types';
-import { DELEGATION_PREFIX, generateEIP7702BatchTransaction } from './eip7702';
+import {
+  DELEGATION_PREFIX,
+  doesAccountSupportEIP7702,
+  generateEIP7702BatchTransaction,
+} from './eip7702';
 import { getGasEstimateBuffer, getGasEstimateFallback } from './feature-flags';
 import { getChainId, rpcRequest } from './provider';
 
@@ -216,7 +220,12 @@ export async function estimateGasBatch({
     chainIds: [chainId],
   });
 
-  const chainResult = is7702Result.find((result) => result.chainId === chainId);
+  const accountSupports7702 = doesAccountSupportEIP7702(messenger, from);
+
+  const chainResult = accountSupports7702
+    ? is7702Result.find((result) => result.chainId === chainId)
+    : undefined;
+
   const isUpgradeRequired = Boolean(chainResult && !chainResult.isSupported);
 
   if (isUpgradeRequired && !chainResult?.upgradeContractAddress) {